Move classes in rules.h/cpp to separate header/implementation files.
