+++ python/setup.py
@@ -27,6 +27,7 @@ sources = [
     'boomer/common/cpp/input/label_matrix_dense.cpp',
     'boomer/common/cpp/input/label_matrix_dok.cpp',
     'boomer/common/cpp/input/nominal_feature_mask_dok.cpp',
+    'boomer/common/cpp/model/condition_list.cpp',
     'boomer/common/cpp/sampling/random.cpp',
     'boomer/common/cpp/sampling/feature_sampling_no.cpp',
     'boomer/common/cpp/sampling/feature_sampling_random.cpp',
@@ -47,7 +48,6 @@ sources = [
     'boomer/common/cpp/binning/feature_binning_equal_frequency.cpp',
     'boomer/common/cpp/binning/feature_binning_equal_width.cpp',
     'boomer/common/cpp/predictions.cpp',
-    'boomer/common/cpp/rules.cpp',
     'boomer/common/cpp/rule_evaluation.cpp',
     'boomer/common/cpp/statistics.cpp',
     'boomer/common/cpp/thresholds.cpp',
